Export elfutils headers and lose the version number from the path.
Having the version number in the path just means that diffs for updates
are completely unusable.

perf refers to libdw.h, libdwfl.h, and version.h as <elfutils/libdw.h> and
so on, so we do still need to do some manual work there, but let's use
symlinks rather than duplicating the header files.

Change-Id: I9fd3c5f5024bde12747fdb1c06d21cdcb3418f03