Add build description for libdw
This adds build description for more elfutils libraries in order support
users that require libdw, such as pahole and libabigail.

While the remaining libraries could be considered standalone, they are
not really suitable to be built standalone in soong as they have
circular dependencies and are usually anyway built as one unit. Hence
only build libdw and include all libraries into it. This then works as
an extension for libelf.

The generated files in libcpu are now all generated from sources. No
really significant generated file is now checked in. Left are config.h
and version.h, which are generated at update time via post_update.sh.

The build rules for the generated files have been reverse engineered
from the autotools build. This mostly making m4, bison and flex work
what we want them to, but builtin rules in soong made that possible
without being overly verbose. An additional library - libeu - needed to
be added as a generator dependency. In order to not overload the main
Android.bp with all the little generator details, subpackages have been
reintroduced and visibility adjusted accordingly.

Bug: 159682555
Change-Id: Ie6f0713812f10b5b12f4e0bdfca27d0369bb71a4
Signed-off-by: Matthias Maennich <maennich@google.com>