elfutils: refactor build description
Refactor the build description to maintain a set of default flags and
options that will also be adopted when adding more elfutils libraries to
this project. Also, move the definition of libelf to the top level as
the subpackage is introducing unnecessary complications with visibility.
While doing so, replace the explicit definition of source files by a
glob in order to include all source files even if this list changes as
part of an update.

Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I2bbfa1a154c71d0c1b05a25a1d8bf85ba2459147