lib: remove usage of `sys/cdefs.h`
This header is a BSD header that is also available in glibc.  However,
this is a not a standard C header and was used for `__CONCAT`.  Because
this is not a standard header, not all libc implementations provide the
header.  Remove the usage of the header and always use the previously
fallback path.  This is needed in order to build with musl.

Signed-off-by: Saleem Abdulrasool <abdulras@google.com>
(cherry picked from commit d390548df1942e98a1d836269a5e41ba52e121f1)
Change-Id: I3284735f88501f8c03549eed3888397d202e3a5f