report_r_debug: handle `-z separate-code' and find more modules
Looking at some cores in eu-stack, I found that they were not being
backtraced.

This was because elfutils had not found some modules (e.g. libc-2.22.so)
in report_r_debug.

That is because it has a limit on the number of link map entries it will
look at, to avoid loops in corrupted core files.

The example I found had:
- 36 elements
- 109 iterations

See also discussion here:

    https://sourceware.org/pipermail/elfutils-devel/2023q2/006149.html

Signed-off-by: Luke Diamand <ldiamand@roku.com>